bug fix
- storage deleted when code run in testing modality (crossref_process.py, datacite_process.py, datacite_process.py, openaire_process.py)
- poetry.lock update
- test bug fix in crossref_process_test.py, datacite_process_test.py, jalc_process_test.py, openaire_process_test.py,  pubmed_processing_test.py)
- test commented in get_pub_test.py because of DataCite API 500 Internal Server Error

+++ test/crossref_process_test.py
@@ -116,7 +116,8 @@ class CrossrefProcessTest(unittest.TestCase):
         self.assertTrue(citations_files_n == 1)
 
         shutil.rmtree(self.output)
-        os.remove(self.db)
+        if os.path.exists(self.db):
+            os.remove(self.db)
 
     def test_preprocess_base_and_decompress_with_cited_redis(self):
         """CASE2: compressed input with cited entities"""
@@ -215,8 +216,8 @@ class CrossrefProcessTest(unittest.TestCase):
         shutil.rmtree(self.output)
         shutil.rmtree(citations_output_path)
 
-
-        os.remove(self.db)
+        if os.path.exists(self.db):
+            os.remove(self.db)
 
     def test_any_db_creation_redis_no_testing(self):
         try:
